+++ example/lib/app.dart
@@ -1,5 +1,7 @@
 import 'package:example/pages/authentication/authentication_page.dart';
 import 'package:example/pages/even_odd/even_odd_page.dart';
+import 'package:example/pages/events/events_page.dart';
+import 'package:example/pages/events_and_builders/events_and_builders_page.dart';
 import 'package:example/pages/fibonnacci/fibonacci_page.dart';
 import 'package:example/pages/counter/counter_page.dart';
 import 'package:example/pages/home/home.dart';
@@ -28,6 +30,8 @@ class MyApp extends StatelessWidget {
         '/fibonnacci': (context) => const FibonnacciPage(),
         '/authentication': (context) => const AuthenticationPage(),
         '/evenOdd': (context) => const EvenOddPage(),
+        '/events': (context) => const EventsPage(),
+        '/eventsAndBuilders': (context) => const EventsAndBuildersPage(),
       },
       initialRoute: initialRoute,
     );