+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * create watchers
 * create states
 * create builders
+* create events